+++ src/main/java/com/soyeon/nubim/domain/user_block/UserBlockService.java
@@ -12,10 +12,7 @@ import com.soyeon.nubim.domain.user_block.dto.UserBlockCreateResponse;
 import com.soyeon.nubim.domain.user_block.dto.UserBlockDeleteResponse;
 import com.soyeon.nubim.domain.user_block.dto.UserBlockReadResponse;
 import com.soyeon.nubim.domain.user_block.dto.UserBlockRequest;
-import com.soyeon.nubim.domain.user_block.exception.AlreadyBlockedException;
-import com.soyeon.nubim.domain.user_block.exception.BlockedUserAccessDeniedException;
 import com.soyeon.nubim.domain.user_block.exception.MultipleUserBlockDeletedException;
-import com.soyeon.nubim.domain.user_block.exception.SelfBlockException;
 import com.soyeon.nubim.domain.user_block.exception.UserBlockDeleteFailException;
 import com.soyeon.nubim.domain.userfollow.UserFollowService;
 
@@ -31,6 +28,7 @@ public class UserBlockService {
 	private final UserService userService;
 	private final UserBlockRepository userBlockRepository;
 	private final UserBlockMapper userBlockMapper;
+	private final UserBlockValidator userBlockValidator;
 	private final UserFollowService userFollowService;
 
 	@Transactional
@@ -39,8 +37,8 @@ public class UserBlockService {
 		User blockingUser = new User(currentUserId);
 		User blockedUser = userService.findByNickname(userBlockRequest.getBlockedUserNickname());
 
-		checkSelfBlock(blockingUser, blockedUser);
-		validateUserBlockNotExists(blockingUser, blockedUser);
+		userBlockValidator.checkSelfBlock(blockingUser, blockedUser);
+		userBlockValidator.validateUserBlockNotExists(blockingUser, blockedUser);
 
 		UserBlock userBlock = userBlockMapper.toEntity(blockingUser, blockedUser);
 		UserBlock savedUserBlock = userBlockRepository.save(userBlock);
@@ -75,21 +73,4 @@ public class UserBlockService {
 		throw new MultipleUserBlockDeletedException();
 	}
 
-	public void checkBlockRelation(User currentUser, User targetUser) {
-		if (userBlockRepository.existsBlockRelationBetweenUser(currentUser, targetUser)) {
-			throw new BlockedUserAccessDeniedException();
-		}
-	}
-
-	private void checkSelfBlock(User blockingUser, User blockedUser) {
-		if (blockingUser.getUserId().equals(blockedUser.getUserId())) {
-			throw new SelfBlockException();
-		}
-	}
-
-	private void validateUserBlockNotExists(User blockingUser, User blockedUser) {
-		if (userBlockRepository.existsByBlockingUserAndBlockedUser(blockingUser, blockedUser)) {
-			throw new AlreadyBlockedException();
-		}
-	}
 }
