+++ src/main/java/com/soyeon/nubim/domain/userfollow/UserFollowService.java
@@ -9,6 +9,7 @@ import com.soyeon.nubim.domain.user.User;
 import com.soyeon.nubim.domain.user.UserMapper;
 import com.soyeon.nubim.domain.user.UserService;
 import com.soyeon.nubim.domain.user.dto.UserSimpleResponseDto;
+import com.soyeon.nubim.domain.user_block.UserBlockValidator;
 import com.soyeon.nubim.domain.userfollow.dto.FollowUserResponseDto;
 import com.soyeon.nubim.domain.userfollow.exception.FollowingStatusException;
 
@@ -21,6 +22,7 @@ public class UserFollowService {
 	private final UserService userService;
 	private final UserMapper userMapper;
 	private final LoggedInUserService loggedInUserService;
+	private final UserBlockValidator userBlockValidator;
 
 	public FollowUserResponseDto createFollow(String nickname) {
 		User followee = userService.getUserByNickname(nickname);
@@ -33,6 +35,7 @@ public class UserFollowService {
 		if (this.isFollowing(follower, followee)) {
 			throw FollowingStatusException.alreadyFollowing(nickname);
 		}
+		userBlockValidator.checkBlockRelation(follower, followee);
 
 		UserFollow userFollow = UserFollow.builder()
 			.follower(follower)
